Create CLI & pre-commit
