+++ no_optional/__main__.py
@@ -1,0 +1,4 @@
+from no_optional.main import app
+
+if __name__ == "__main__":
+    app()
