+++ no_optional/main.py
@@ -1,71 +1,41 @@
-import libcst as cst
-from libcst import matchers as m
-from libcst.codemod import VisitorBasedCodemodCommand
+import contextlib
+import time
+from os import devnull
+from pathlib import Path
+from typing import List
+
+from libcst.codemod import CodemodContext, parallel_exec_transform_with_prettyprint
+from typer import Argument, Exit, Typer, secho
+
+from no_optional import NoOptionalCommand
+
+app = Typer()
 
 
-class NoOptionalCommand(VisitorBasedCodemodCommand):
-    @m.leave(
-        m.Subscript(
-            value=m.Name(value="Optional")
-            | m.Attribute(value=m.Name(value="typing"), attr=m.Name(value="Optional"))
-        )
-    )
-    def replace_optional(
-        self, original_node: cst.Subscript, updated_node: cst.Subscript
-    ) -> cst.Subscript:
-        if original_node.value.value == "Optional":
-            return updated_node.with_changes(
-                value=cst.Name("Union"),
-                slice=[
-                    *updated_node.slice,
-                    cst.SubscriptElement(
-                        slice=cst.Index(value=cst.Name("None")),
-                        comma=cst.MaybeSentinel.DEFAULT,
-                    ),
-                ],
-            )
-        else:
-            return updated_node.with_changes(
-                value=cst.Attribute(value=cst.Name("typing"), attr=cst.Name("Union")),
-                slice=[
-                    *updated_node.slice,
-                    cst.SubscriptElement(
-                        slice=cst.Index(value=cst.Name("None")),
-                        comma=cst.MaybeSentinel.DEFAULT,
-                    ),
-                ],
+@app.command()
+def command(
+    files: List[Path] = Argument(..., exists=True, dir_okay=True, allow_dash=True)
+):
+    transformer = NoOptionalCommand(CodemodContext())
+    start_time = time.time()
+
+    with open(devnull, "w") as null:
+        with contextlib.redirect_stderr(null):
+            result = parallel_exec_transform_with_prettyprint(
+                transformer,
+                [str(file) for file in files],
+                include_generated=True,
+                show_successes=True,
+                jobs=1,
             )
 
-    @m.call_if_inside(
-        m.Annotation(
-            annotation=m.Subscript(
-                value=m.Name(value="Optional")
-                | m.Attribute(value=m.Name("typing"), attr=m.Name("Optional")),
-                slice=(
-                    m.SubscriptElement(
-                        slice=m.Index(value=m.Subscript(value=m.Name(value="Union"))),
-                    ),
-                    m.ZeroOrMore(),
-                ),
-            )
-        )
-    )
-    @m.leave(
-        m.Subscript(
-            value=m.Name(value="Optional")
-            | m.Attribute(value=m.Name("typing"), attr=m.Name("Optional"))
-        )
-    )
-    def remove_union_redundancy(
-        self, original_node: cst.Subscript, updated_node: cst.Subscript
-    ) -> cst.Subscript:
-        return updated_node.with_changes(
-            slice=[*updated_node.slice[0].slice.value.slice, *updated_node.slice[1:]]
-        )
+    modified = [f for f in files if f.stat().st_mtime > start_time]
+    for file in modified:
+        secho(f"refactored {file.relative_to('.')}", bold=True)
 
-    @m.call_if_inside(m.ImportAlias(name=m.Name(value="Optional")))
-    @m.leave(m.Name(value="Optional"))
-    def replace_import(
-        self, original_node: cst.Name, updated_node: cst.Name
-    ) -> cst.Name:
-        return updated_node.with_changes(value="Union")
+    if modified:
+        raise Exit(1)
+
+
+if __name__ == "__main__":
+    app()
