+++ setup.cfg
@@ -28,12 +28,17 @@ packages = find:
 include_package_data = True
 install_requires =
     libcst
+    typer
 python_requires = >=3.7
 
 [options.extras_require]
 test =
     coverage[toml] >= 6.2
     pytest >= 6.2.5
+
+[options.entry_points]
+console_scripts =
+    no-optional = no_optional.main:app
 
 [flake8]
 statistics = True
